+++ database/migrations/2022_10_28_092618_create_activity_log_table.php
@@ -26,16 +26,24 @@ use Illuminate\Support\Facades\Schema;
 return new class extends Migration {
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->timestamp('last_active_at')->nullable();
-            $table->unsignedTinyInteger('invalid_login_attempts')->default(0);
-        });
+        Schema::connection(config('activitylog.database_connection'))
+            ->create(config('activitylog.table_name'), function (Blueprint $table) {
+                $table->bigIncrements('id');
+                $table->string('log_name')->nullable();
+                $table->text('description');
+                $table->nullableMorphs('subject', 'subject');
+                $table->string('event')->nullable();
+                $table->nullableMorphs('causer', 'causer');
+                $table->json('properties')->nullable();
+                $table->uuid('batch_uuid')->nullable();
+                $table->timestamps();
+                $table->index('log_name');
+            });
     }
 
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn(['last_active_at', 'invalid_login_attempts']);
-        });
+        Schema::connection(config('activitylog.database_connection'))
+            ->dropIfExists(config('activitylog.table_name'));
     }
 };