+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deleting referral links
 - Referral links pagination
 - Server events log
+- Logging changes in users table
 ### Changed
 - Network host which does not broadcast is marked as failed before deleting
 - Do not request banners when zones are invalid